+++ integrations.go
@@ -3,6 +3,10 @@ package extensions
 import (
 	"github.com/gosimple/slug"
 	"github.com/wakflo/extensions/internal/integrations/calendly"
+	"github.com/wakflo/extensions/internal/integrations/discord"
+	"github.com/wakflo/extensions/internal/integrations/docconverter"
+	"github.com/wakflo/extensions/internal/integrations/telegrambot"
+	"github.com/wakflo/extensions/internal/integrations/whatsapp"
 
 	// "github.com/wakflo/extensions/internal/integrations/csv"
 
@@ -124,6 +128,10 @@ func RegisterIntegrations() sdk.IntegrationsRegistrar {
 		zohosalesiq.Integration,       // ZohoSales iq
 		surveyMonkey.Integration,      // Survey Monkey
 		pinterest.Integration,         // Pinterest
+		discord.Integration,           // Discord
+		docconverter.Integration,      // Doc Converter
+		whatsapp.Integration,          // Whatsapp Business
+		telegrambot.Integration,       // Telegram Bot
 	}
 
 	// 🛑Do-Not-Edit