+++ internal/integrations/googlemail/actions/doc.go
@@ -16,3 +16,9 @@ var sendEmailDocs string
 
 //go:embed send_email_template.md
 var sendEmailTemplateDocs string
+
+//go:embed extract_email_text.md
+var extractMailTextDocs string
+
+//go:embed extract_document_from_email.md
+var extractDocumentsFromEmailDocs string